Attempt to create lab 2 script
An attempt to create a simple script to run the robot through desired poses using rostopic pub to directly publish joint values to the robot. Only works with the hardware due to publishing joint states directly to the motors. The recommended implementation of this would be to use MoveIt's motion planning API, but I did not have time to learn it. I also created this directly before I stopped being able to connect to the robot using the hardware setup script so it may just be nonfunctional.